+++ crates/node/src/pruning.rs
@@ -541,6 +541,7 @@ mod tests {
                         &block,
                         &ChainConfig {
                             log_address_activation_height: None,
+                            algorithm_voting_window: None,
                             algorithm_proposal_staging_height: None,
                             algorithm_quorum_activation_height: None,
                             algorithm_timelock_activation_height: None,
@@ -766,6 +767,7 @@ mod tests {
                         &block,
                         &ChainConfig {
                             log_address_activation_height: None,
+                            algorithm_voting_window: None,
                             algorithm_proposal_staging_height: None,
                             algorithm_quorum_activation_height: None,
                             algorithm_timelock_activation_height: None,